Extract common test classes outside RecyclerView
